+++ cmd/cli/main.go
@@ -39,7 +39,38 @@ func main() {
 	playersRepository := players.NewPlayerRepository(dbHandler)
 	playersManager := players.NewManager(playersRepository)
 
-	if err := commands.NewRootCommand(playersManager, gamesManager, seasonManager).Execute(); err != nil {
+	rootCommand := createCommands(seasonManager, gamesManager, playersManager)
+
+	if err := rootCommand.Execute(); err != nil {
 		log.Fatal(err)
 	}
+}
+
+func createCommands(seasonsManager seasons.Manager, gamesManager games.Manager, playersManager players.Manager) commands.Command {
+	createPlayer := commands.NewCreatePlayerCommand(playersManager)
+	getPlayers := commands.NewGetPlayersCommand(gamesManager)
+	playersCommand := commands.NewPlayersCommand()
+	playersCommand.AddCommand(createPlayer)
+	playersCommand.AddCommand(getPlayers)
+
+	createSeason := commands.NewCreateSeasonCommand(seasonsManager)
+	getSeason := commands.NewGetSeasonsCommand(seasonsManager)
+	activateSeason := commands.NewActivateSeasonCommand(seasonsManager)
+	tableCommand := commands.NewGetTableCommand(gamesManager, seasonsManager)
+	seasonsCommand := commands.NewSeasonsCommand()
+	seasonsCommand.AddCommand(createSeason)
+	seasonsCommand.AddCommand(getSeason)
+	seasonsCommand.AddCommand(activateSeason)
+	seasonsCommand.AddCommand(tableCommand)
+
+	createGame := commands.NewCreateGameCommand(gamesManager, playersManager)
+	gamesCommands := commands.NewGamesCommand()
+	gamesCommands.AddCommand(createGame)
+
+	rootCommand := commands.NewRootCommand()
+	rootCommand.AddCommand(playersCommand)
+	rootCommand.AddCommand(seasonsCommand)
+	rootCommand.AddCommand(gamesCommands)
+
+	return rootCommand
 }
