+++ internal/cli/commands/root.go
@@ -2,12 +2,6 @@ package commands
 
 import (
 	"github.com/spf13/cobra"
-
-	"github.com/spie/fskick/internal/cli/commands/games"
-	"github.com/spie/fskick/internal/cli/commands/seasons"
-	g "github.com/spie/fskick/internal/games"
-	"github.com/spie/fskick/internal/players"
-	s "github.com/spie/fskick/internal/seasons"
 )
 
 type Command interface {
@@ -19,19 +13,12 @@ type rootCommand struct {
 	cc *cobra.Command
 }
 
-func NewRootCommand(playersManager players.Manager, gamesManager g.Manager, seasonsManagers s.Manager) *rootCommand {
+func NewRootCommand() *rootCommand {
 	rootCommand := rootCommand{cc: &cobra.Command{
 		Use:   "fskick",
 		Short: "FSKick CLI app",
 		Long:  "CLI app for FSKick to create new players, seasons, games and show results and statistics",
 	}}
-
-	playersCommand := NewPlayersCommand(playersManager, gamesManager)
-	rootCommand.AddCommand(playersCommand)
-	seasonsCommand := seasons.NewSeasonsCommand(gamesManager, seasonsManagers)
-	rootCommand.cc.AddCommand(seasonsCommand.GetSeasonsCommand())
-	gamesCommand := games.NewGamesCommand(gamesManager, playersManager)
-	rootCommand.cc.AddCommand(gamesCommand.GetGamesCommand())
 
 	return &rootCommand
 }
@@ -42,4 +29,8 @@ func (command *rootCommand) Execute() error {
 
 func (command *rootCommand) AddCommand(c Command) {
 	command.cc.AddCommand(c.getCommand())
+}
+
+func (command *rootCommand) getCommand() *cobra.Command {
+	return command.cc
 }
